Clean up
We made a couple changes:
* Changed packages for a couple classes to prevent package dependency
cycles, and to separate API from implementations
* Removed a second OpInstance implementation, not sure how two
implementations came about :P
