+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/AbstractOpTest.java
@@ -50,9 +50,9 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.scijava.Context;
 import org.scijava.cache.CacheService;
-import org.scijava.ops.OpHistoryService;
 import org.scijava.ops.OpService;
 import org.scijava.ops.core.Op;
+import org.scijava.ops.provenance.OpHistoryService;
 import org.scijava.plugin.PluginService;
 import org.scijava.thread.ThreadService;
 import org.scijava.types.TypeService;
