+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/coloc/ColocalisationTest.java
@@ -59,9 +59,9 @@ import org.scijava.Context;
 import org.scijava.app.StatusService;
 import org.scijava.cache.CacheService;
 import org.scijava.io.location.FileLocation;
-import org.scijava.ops.OpHistoryService;
 import org.scijava.ops.OpService;
 import org.scijava.ops.core.builder.OpBuilder;
+import org.scijava.ops.provenance.OpHistoryService;
 import org.scijava.plugin.PluginService;
 import org.scijava.thread.ThreadService;
 import org.scijava.types.TypeService;
