+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/filter/vesselness/FrangiVesselnessTest.java
@@ -48,9 +48,9 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.scijava.Context;
 import org.scijava.cache.CacheService;
-import org.scijava.ops.OpHistoryService;
 import org.scijava.ops.OpService;
 import org.scijava.ops.core.builder.OpBuilder;
+import org.scijava.ops.provenance.OpHistoryService;
 import org.scijava.plugin.PluginService;
 import org.scijava.script.ScriptService;
 import org.scijava.thread.ThreadService;
