+++ scijava/scijava-ops/src/test/java/org/scijava/ops/provenance/ProvenanceTest.java
@@ -15,7 +15,6 @@ import org.scijava.function.Producer;
 import org.scijava.ops.AbstractTestEnvironment;
 import org.scijava.ops.OpDependency;
 import org.scijava.ops.OpField;
-import org.scijava.ops.OpHistoryService;
 import org.scijava.ops.OpInfo;
 import org.scijava.ops.OpMethod;
 import org.scijava.ops.core.OpCollection;
