+++ scijava/scijava-ops/templates/main/java/org/scijava/ops/matcher/OpWrappers.vm
@@ -19,7 +19,7 @@ import org.scijava.ops.OpInfo;
 import org.scijava.ops.hints.BaseOpHints.DependencyMatching;
 import org.scijava.ops.hints.Hints;
 import org.scijava.ops.provenance.OpExecutionSummary;
-import org.scijava.ops.OpHistoryService;
+import org.scijava.ops.provenance.OpHistoryService;
 import org.scijava.ops.util.OpWrapper;
 import org.scijava.plugin.Plugin;
 import org.scijava.types.GenericTyped;
